Finish up initial implementation and version for packaging
