+++ src/ElectricityNetworFileParser/dataclasses.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+@dataclass
+class PropertyDescription:
+    property_type : str
+    property_attributes : dict
