--- src/NetworFileParser/dataclasses.py
@@ -1,6 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class PropertyDescription:
-    property_type : str
-    property_attributes : dict
